added authentication middleware to meals route

+++ controllers/mealsController.ts
@@ -4,14 +4,19 @@ import { StatusCodes } from "http-status-codes";
 import meals from "../models/meals";
 
 export const createMeals = async (req: Request, res: Response) => {
+  const { id } = req.user;
+  if (!id) {
+    res.status(404).json({ message: "Not authorized" });
+    return;
+  }
   try {
-    const { name, description, amount, price, admin } = req.body;
+    const { name, description, amount, price } = req.body;
     const mealsCreated = await meals.create({
       name,
       description,
       amount,
       price,
-      admin,
+      admin: id,
     });
     mealsCreated
       ? res
@@ -49,9 +54,10 @@ export const getOneMeal = async (req: Request, res: Response) => {
 export const updateMeal = async (req: Request, res: Response) => {
   try {
     const { id } = req.params;
+    const { id: admin } = req.user;
     const updateBody = req.body;
 
-    const meal = await mealsDb.findOne({ _id: id });
+    const meal = await mealsDb.findOne({ admin });
     if (!meal) {
       res.status(404).json({ message: "meal not found" });
       return;
@@ -70,6 +76,13 @@ export const updateMeal = async (req: Request, res: Response) => {
 export const deleteMeals = async (req: Request, res: Response) => {
   try {
     const { id } = req.params;
+    const { id: admin } = req.user;
+
+    const meal = await mealsDb.findOne({ admin });
+    if (!meal) {
+      res.status(404).json({ message: "Not authorized" });
+      return;
+    }
     const deleted = await mealsDb.findByIdAndDelete({ _id: id });
     deleted
       ? res.status(200).json({ message: "Meal deleted successfully" })
